🔨 Rename ambigous component name
